Add regex validation to Telemetry events
This PR adds validation to Telemetry events by adding the regex pattern
to be matched for the following event fields: category, method, object.
The pattern to be matched was defined in Bug 1337022 at the link
https://bugzilla.mozilla.org/show_bug.cgi?id=1337022
This PR also adds the events found in the doc
https://gecko.readthedocs.io/en/latest/toolkit/components/telemetry/telemetry/collection/events.html
to the validation test for the main schema.