Fix to the gc schema to handle variations
The schema of the "gc" section landed in bug 1308040, and was
subsequently modified in bugs 1386511 and 1412729. Update the
"gc" schema to handle these variations.

Initially, the array field was called `slices`, then it changed to `slices_list`
and reused the `slices` field for what was previously called `num_slices`.

The updated schema allows both. Further, the maximum number of items
for various parts of the gc data changed, so we use the high water mark 
for the max in the schema.