Bug 1435827: Adding JSON and Parquet schemas (#215)
* Bug 1435827: Adding JSON and Parquet schemas

This adds the necessary validation and ingestion schemas for the new "Untrusted Modules" ping introduced in Bug 1435827, along with basic test payloads.

* Bug 1435827: Fixing comment syntax

Now using "description" and "examples" instead of "$comment" for JSON schemas.

* Bug 1435827: Fixing required / optional fields in the environment block

Some fields in the environment block were marked as required when the JSON schema does not require them. Fixed to synchronize the schemas.