Disallow camelCase names for document types
As part of this change, we provide symlinks for compliant kebab-case names
for the two existing docTypes which are camelCase. Both variants produce
output in the schemas/ directory and mozilla-schema-generator will be updated
to ignore the old names when creating BQ schemas.

My expectation is that this change to the schema deployment pipeline will
be effectively invisible to the AWS and GCP pipelines. They should both
continue functioning correctly without changes.