Remove most parquet schemas
With teardown of the AWS pipeline, direct2parquet no longer exists,
meaning the Parquet schemas in this project are no longer useful.
This PR removes them and updates the instructions for creating new schemas
to skip any mention of Parquet schemas.

We retain the shield-study and pioneer parquet schemas here, as they follow
some different rules and may still be somewhat relevant.