Allow negative sessionLengths and event timestamps
These values should never be negative, but we are seeing a significant number
of negative values since Firefox 71. Since we have not been able to determine
the cause and the current situation is leading us to undercount in all desktop
analyses, we are going to allow negative values to start flowing through.

See bugs
https://bugzilla.mozilla.org/show_bug.cgi?id=1592012#c15
and
https://bugzilla.mozilla.org/show_bug.cgi?id=1602521