AET update: 64-char user_ids and relaxing validation
Fixes #571

Also allows `firefox-accounts` pings without an `ecosystem_client_id` value as
long as `ecosystem_anon_id` is present. Server apps will need to send AET
pings with anon_id but without any relevant client_id to be associated with,
so we want to avoid having to introduce a static server-side client_id just to
conform to the schema.

Telemetry AET pings are still required to have an `ecosystem_client_id` as we
expect the client should always be able to provide that value.