Bug 1654558 Add xfocsp-error-report schema
Closes https://github.com/mozilla-services/mozilla-pipeline-schemas/pull/570

This pulls in the commit from that PR and duplicates it in a new namespace.
We will be rerouting these pings in the pipeline so that they can have
stricter access controls.

There is a test case added here so that CI will fail if these two schemas
get out of sync.