Add common ping fields to xfocsp-error-report
In testing, we found several top-level fields included in the validation
document were going to additional_properties. We add them here.

Of note, we do not add client_id or environment, which are part of the common
ping format but specifically excluded here for user privacy.

Reverts https://github.com/mozilla-services/mozilla-pipeline-schemas/pull/582