Require JSON Pointer format for jwe_mappings
This moves us from using an informal ad hoc format for specifying paths to
a proposed IETF standard.

I came across this standard as I was starting to draft a PR for relying on
jwe_mappings in the AET decoder. In particular, the Jackson library provides
an implementation of JSON Pointer that simplifies the AET logic.

See draft: https://github.com/mozilla/gcp-ingestion/pull/1374