Bug 1684971 - Add previous notification action to default browser agent telemetry (#652)
* Bug 1621703 - Add default browser notification telemetry

* Bug 1621703 - Remove default browser agent's client id

Originally, we planned to have a client id in the default browser
agent's ping. However, we eventually decided not to include it. We never
merged the code that added the client id, so no pings have ever been
sent with this field.

* Bug 1621703 - Deprecate default browser agent's client id

Originally, we planned to have a client id in the default browser
agent's ping. However, we eventually decided not to include it. We never
merged the code that added the client id, so no pings have ever been
sent with this field.

* Bug 1684971 - Add previous notification action field

Add previous notification action field to the default browser agent ping

Co-authored-by: bytesized <bytesized@mozilla.com>