Bug 1751739 - Add enrollment_id to Glean experiments
Glean experiment extras _should_ be `MAP<STRING, STRING>`
types. Instead, they are incorrectly typed as a
`STRUCT<type STRING>`.

With this change, we will accomodate the `enrollmentId`
field by updating the `extra` field to by:
`STRUCT<type STRING, enrollment_id STRING>`.

This should allow the `enrollmentId`s to propogate,
and the experiments should not be included in
`additionalProperties` any longer.