Bug 1901263: Add `profileGroupId` to schemas.
This adds the `profileGroupId` added to the client side in bug 1901263
to the schemas wherever the `clientId` is present. I believe that every
ping including `clientId` will include `profileGroupId` however since
the new property is not marked as required in any schema (as it cannot
be to avoid breaking older versions of Firefox) if new pings are
introduced that do not include the property then it should not break
anything.